Ametsuchi: add SpecificQueryExecutor interface
