+++ irohad/ametsuchi/impl/command_executor.cpp
@@ -1,3 +1,8 @@
+/**
+ * Copyright Soramitsu Co., Ltd. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 #include "ametsuchi/command_executor.hpp"
 
 #include "interfaces/commands/command.hpp"
