+++ irohad/ametsuchi/impl/tx_executor.cpp
@@ -1,3 +1,8 @@
+/**
+ * Copyright Soramitsu Co., Ltd. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 #include "ametsuchi/tx_executor.hpp"
 
 #include "interfaces/commands/command.hpp"
